Split text-mode-editor from text-mode DSM plugin
