+++ package.json
@@ -16,6 +16,8 @@
     "#parsing/*": "./parsing/*",
     "#tests": "./src/tests/puppeteer-utils.ts",
     "#metadata/*": "./metadata/*",
+    "#text-mode-core": "./text-mode-core/index.ts",
+    "#text-mode-editor": "./text-mode-editor/index.ts",
     "#components": "./src/components/index.ts"
   },
   "eslintIgnore": [
