--- src/plugins/text-mode/docs/intro.md
@@ -1,3 +1,0 @@
-# DesModder Text Mode Docs
-
-This documentation has moved! Visit https://www.desmodder.com/text-mode.
