+++ text-mode-editor/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../tsconfig-base.json",
+  "references": [{ "path": "../text-mode-core" }],
+  "include": ["**/*"]
+}
