+++ text-mode-editor/view/editor.less
