+++ text-mode-editor/view/plugins/StyleCircle.less
