+++ text-mode-editor/view/plugins/StyleCircle.tsx
