+++ text-mode-editor/view/plugins/checkboxWidget.ts
