+++ text-mode-editor/view/plugins/collapseStylesAtStart.ts
