+++ text-mode-editor/view/plugins/footerWidget.less
