+++ text-mode-editor/view/plugins/footerWidget.ts
