+++ text-mode-editor/view/plugins/styleCircles.ts
