+++ text-mode-editor/view/plugins/styleMappingWidgets/imageWidget.tsx
