+++ text-mode-editor/view/plugins/styleMappingWidgets/index.ts
