+++ text-mode-editor/view/plugins/styleMappingWidgets/inlineToggleWidget.less
