+++ text-mode-editor/view/plugins/styleMappingWidgets/inlineToggleWidget.tsx
