+++ tsconfig.json
@@ -7,6 +7,7 @@
     { "path": "./text-mode-core" },
     { "path": "./localization" },
     { "path": "./parsing" },
+    { "path": "./text-mode-editor" },
     { "path": "./src" }
   ]
 }
